Make this repo a working, installable package
Without the additional requirements and the data folder being included
in the package, you can't just add this repo to another project's
requirements.txt and install it, you have to clone it and install the
requirements.txt, then install the folder as a module itself.

I couldn't see any reason to exclude these packages from setup.py, as
the versions we need are released and available on PyPi, so I added them
with the same minimumn version constraints as in requirements.in.

Likewise, the schema was already bundled in the repo, I've just moved it
to a different location so that it can be in the package too. I agree
this is less desirable in the long run, but since the schema is very
slow moving at the moment, I don't think it's a huge issue.