Document how to run tests in README.md
I had to google it and guess from pytests inclusion in the
requirements_dev, so to save the next person the trouble :).

Looks like there were a few trailing spaces which my editor
auto-removed. Normally I'd split that into a separate commit, but since
this file is so small I don't think it's too confusing.